fix(notifications): serve org stats via JWT private endpoint
Move org dashboard notification aggregation to
/private/org_notification_stats with middlewareAuth, harden days
parsing, add missing not-authenticated copy, and cover new query
validation paths in unit tests.
